fix(ts build): update typescript typings to remove false positives in vscode problems window
The build output has  cjs and esm, the default is esm in the tsconfig.json which also excluded tests
so they weren’t included in the output. This meant errors were shown in vscode because the
tsconfig.json is used to report problems. Both the cjs and esm builds are now handled in separate
files and the tsconfig is used for development.